+++ docs/components/custom/SakuraSidebarCustom.vue
@@ -1,6 +1,0 @@
-<template>
-  <SakuraSidebar>
-    <!-- TODO:  -->
-    <SidebarThemeDynamic />
-  </SakuraSidebar>
-</template>