+++ docs/components/playground/SakuraHeroBackgroundPG.vue
@@ -1,0 +1,18 @@
+<script setup lang="ts">
+import { ref } from 'vue'
+
+const config = ref({
+  banner: {
+    urls: 'https://wrxinyue-images.s3.bitiful.net/pc-wallpaper/wallhaven-pkxlv3.jpg',
+  },
+})
+
+function configUpdate(newConfig: any) {
+  config.value = newConfig
+}
+</script>
+
+<template>
+  <SakuraHeroBackground :urls="config.banner.urls" />
+  <CodeEditor :config="config" @update:config="configUpdate" />
+</template>