+++ docs/components/playground/components-custom/SakuraArticlePinnedCustomPG.vue
@@ -1,6 +1,0 @@
-<template>
-  <div class="<md:hidden">
-    <SakuraIconTextDivider icon="i-fa6-solid:anchor" text="START:DASH!!" />
-    <PostPinnedThemeStartDash />
-  </div>
-</template>