+++ docs/components/playground/components-custom/SakuraFooterCustomPG.vue
@@ -1,3 +1,0 @@
-<template>
-  <SakuraFooter />
-</template>