+++ docs/components/playground/components-custom/SakuraNoticeBoardCustomPG.vue
@@ -1,3 +1,0 @@
-<template>
-  <SakuraNoticeBoard />
-</template>