+++ docs/components/playground/components-themes/ArticleListThemeCardPG.vue
@@ -1,3 +1,0 @@
-<template>
-  <PostListThemeCard />
-</template>