+++ docs/components/playground/components-themes/ArticleListThemeMinimaPG.vue
@@ -1,3 +1,0 @@
-<template>
-  <PostListThemeMinima />
-</template>