+++ docs/components/playground/components-themes/PaginationThemeValaxyPG.vue
@@ -1,3 +1,0 @@
-<template>
-  <PaginationThemeValaxy :page-size="3" :total="30" />
-</template>